Fixed multiple tooltip displays while in 'quarter', 'month', 'week' display.
To smooth out animations, excess datapoints would be set to the initial value of 'quarter', 'month', or 'week' depending on the current display.

This caused the tooltip at that initial datapoint to show 90, 150, and 173 datapoints (respectively).

Created a callback function for the tooltip to omit the excess datapoints at the initial point of the currently displayed graph.